Add files via upload
We were assigned to help a struggling restaurant with managing their inventory. The program you
created will need to be capable of reading the list of ingredients for each dish, storing the ingredients list for
each dish into a bit vector, and compute data analytics on the ingredients list for each dish to help the
restaurant determine where it will need to make improvements. As well as reading all the ingredients of each
dish, you will have to be able to read the list of ingredients the restaurant has available to use for the dishes
and determine if the creation of the dish is even possible with their current inventory. If the dish cannot be
created, you will need to inform the restaurant/User that the dish is not able to be created due to whatever
ingredient is not available.